requirements.txt
google-genai
requests
edge-tts>=6.1.9
yt-dlp>=2024.12.23
moviepy>=2.1.1
openai-whisper
torch
numpy
pillow
imageio-ffmpeg
arabic-reshaper
python-bidi
python-dotenv
opencv-python
supabase
beautifulsoup4
lxml
playwright
